Fixing - Session Timeout Setting Only Partially Works #80

+++ src-tauri/src/app_preference/mod.rs
@@ -39,7 +39,8 @@ impl RecentFile {
 
 #[derive(Debug, Deserialize)]
 pub(crate) struct PreferenceData {
-    session_timeout: Option<u8>,
+    // In minutes; the settings UI accepts 1-1440 (issue #80)
+    session_timeout: Option<u16>,
     clipboard_timeout: Option<u16>,
     default_entry_category_groupings: Option<String>,
     theme: Option<String>,
